Fix: Navigate to details screen only if item is selected
This commit addresses an issue where the app would navigate to the item details screen even if no new item was selected.

The `MainScreenViewModel.updateSelectedItem` function now checks if the `selectedItemName` is different from the `previousSelectedItem?.name` before proceeding with the UI update and navigation.

Additionally, the `MainActivity` now passes the `navController` to the `MainScreen` composable to enable navigation. The `ContentScreen` within `MainScreen` now utilizes this passed `navController`. A utility composable `ShowContentOrEmpty` has been introduced to simplify conditional rendering based on data availability in `ContentScreen`.

A typo in `ItemListScree` has also been corrected to `ItemListScreen`.