Refactor: Add `tab` dimension to theme and use it for TabRow indicator
This commit introduces a new dimension `tab` to the `AppSize` and `LocalAppSize` within `AppTheme.kt` and `DesignSystem.kt`.

The `tab` dimension is initialized to `2.dp` in `defaultAppSize`.

This new dimension is then utilized in `ItemListScreen.kt` to set the `height` of the `TabRowDefaults.SecondaryIndicator`. Previously, this was a hardcoded value. This change centralizes the tab indicator height within the theme, allowing for easier modification and ensuring consistency.